CORTX-29554: Alerts feature archived, support bundle cleaned up
