+++ archive/agent/api.py
@@ -1,5 +1,5 @@
 # CORTX-CSM: CORTX Management web and CLI interface.
-# Copyright (c) 2020 Seagate Technology LLC and/or its Affiliates
+# Copyright (c) 2022 Seagate Technology LLC and/or its Affiliates
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published
 # by the Free Software Foundation, either version 3 of the License, or
@@ -12,17 +12,18 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
+from csm.common.services import Service
 
-from cortx.utils.data.db.db_provider import DataBaseProvider
-from csm.core.blogic.models.support_bundle import SupportBundleModel
-from cortx.utils.data.access import Query
-from cortx.utils.data.access.filters import Compare
+class AlertHttpNotifyService(Service):
+    def __init__(self):
+        super().__init__()
+        self.unpublished_alerts = set()
 
-class SupportBundleRepository:
-    def __init__(self, storage: DataBaseProvider):
-        self.db = storage
+    def push_unpublished(self):
+        while self.unpublished_alerts:
+            self.handle_alert()
 
-    async def retrieve_all(self, bundle_id) -> [SupportBundleModel]:
-        query = Query().filter_by(Compare(SupportBundleModel.bundle_id, '=',
-                                          bundle_id))
-        return await self.db(SupportBundleModel).get(query)
+    def handle_alert(self, alert):
+        self.unpublished_alerts.add(alert)
+        if CsmRestApi.push(alert):
+            self.unpublished_alerts.discard(alert)
