--- csm/cli/schema/bundle_generate.json
@@ -1,61 +1,0 @@
-{
-  "name": "bundle_generate",
-  "description": "Generate's Support Bundle for all the CORTX System.",
-  "permissions_tag": "create",
-  "need_confirmation": false,
-  "args": [
-    {
-      "flag": "bundle_id",
-      "type": "str",
-      "help": "Support Bundle Id."
-    },
-    {
-      "flag": "comment",
-      "type": "str",
-      "help": "Specified Reason for Generating Support Bundle."
-    },
-    {
-      "flag": "node_name",
-      "type": "str",
-      "help": "Current Node Name."
-    },
-    {
-      "flag": "-c",
-      "dest": "components",
-      "type": "str",
-      "choices": [
-        "all",
-        "csm",
-        "sspl",
-        "s3server",
-        "motr",
-        "hare",
-        "provisioner",
-        "os",
-        "health_map",
-        "manifest",
-        "alerts",
-        "uds",
-        "elasticsearch",
-        "ha",
-        "utils"
-      ],
-      "nargs": "+",
-      "help": "Specific Component for Support Bundle to be Generated."
-    },
-    {
-      "flag": "args",
-      "default": [],
-      "suppress_help": true,
-      "nargs": "?",
-      "help": ""
-    }
-  ],
-  "comm": {
-    "type": "direct",
-    "target": "cortx.utils.support_framework",
-    "method": "init",
-    "class": "ComponentsBundle",
-    "is_static": true
-  }
-}
