--- csm/cli/schema/csm_bundle_generate.json
@@ -1,42 +1,0 @@
-{
-  "name": "csm_bundle_generate",
-  "description": "Csm Bundle Generate",
-  "permissions_tag": "create",
-  "args": [
-    {
-      "flag": "component",
-      "type": "str",
-      "choices":[
-        "csm",
-        "alerts",
-        "uds",
-        "elasticsearch"
-      ],
-      "help": "Specify the Name of component for Generating Bundle."
-    },
-    {
-      "flag": "bundle_id",
-      "type": "str",
-      "help": "Specify the Bundle Id for Generating Bundle."
-    },
-    {
-      "flag": "path",
-      "type": "str",
-      "help": "Specify the Path for Generating Bundle."
-    },
-    {
-      "flag": "args",
-      "default": [],
-      "suppress_help": true,
-      "nargs": "?",
-      "help": ""
-    }
-  ],
-  "comm": {
-    "type": "direct",
-    "target": "csm.cli.support_bundle",
-    "method": "init",
-    "class": "CSMBundle",
-    "is_static": true
-  }
-}
