--- csm/cli/schema/support_bundle.json
@@ -1,114 +1,0 @@
-{
-  "name": "support_bundle",
-  "description": "Support Bundle",
-  "sub_commands": [
-    {
-      "name": "generate",
-      "description": "Generate Support Bundle",
-      "need_confirmation": false,
-      "permissions_tag": "create",
-      "args": [
-        {
-          "flag": "comment",
-          "type": "str",
-          "help": "Specify the Reason for Generating Support Bundle."
-        },
-        {
-          "flag": "--os",
-          "action": "store_const",
-          "help": "Generate Only OS Logs",
-          "default": "false",
-          "const": "true"
-        },
-        {
-          "flag": "-c",
-          "dest": "components",
-          "type": "str",
-          "choices":[
-            "all",
-            "csm",
-            "sspl",
-            "s3server",
-            "motr",
-            "hare",
-            "provisioner",
-            "os",
-            "health_map",
-            "alerts",
-            "uds",
-            "elasticsearch",
-            "ha",
-            "utils"
-          ],
-          "nargs": "+",
-          "help": "Generate Bundle for Specific components defined."
-        },
-        {
-          "flag": "args",
-          "default": [],
-          "suppress_help": true,
-          "nargs": "?",
-          "help": ""
-        }
-      ],
-      "comm": {
-        "type": "direct",
-        "target": "cortx.utils.support_framework",
-        "method": "_generate_bundle",
-        "class": "SupportBundle",
-        "is_static": true
-      }
-    },
-    {
-      "name": "status",
-      "description": "",
-      "need_confirmation": false,
-      "permissions_tag": "list",
-      "args": [
-        {
-          "flag": "bundle_id",
-          "type": "str",
-          "help": "Unique Bundle Id provided while generation."
-        },
-        {
-          "flag": "args",
-          "default": [],
-          "suppress_help": true,
-          "nargs": "?",
-          "help": ""
-        },
-        {
-          "flag": "-f",
-          "dest": "format",
-          "help": "Format of Output",
-          "default": "table",
-          "type": "str",
-          "choices": [
-            "table",
-            "xml",
-            "json"
-          ]
-        }
-      ],
-      "comm": {
-        "type": "direct",
-        "target": "cortx.utils.support_framework",
-        "method": "_get_bundle_status",
-        "class": "SupportBundle",
-        "is_static": true
-      },
-      "output": {
-        "table": {
-          "headers": {
-            "bundle_id": "Bundle Id",
-            "comment": "Comment",
-            "node_name": "Node Name",
-            "message": "Message",
-            "result": "Result"
-          },
-          "filters": "status"
-        }
-      }
-    }
-  ]
-}
