+++ csm/cli/support_bundle/__init__.py
@@ -12,7 +12,3 @@
 # along with this program. If not, see <https://www.gnu.org/licenses/>.
 # For any questions about this software or licensing,
 # please email opensource@seagate.com or cortx-questions@seagate.com.
-
-from .csm_bundle_generate import CSMBundle
-from .support_bundle import SupportBundle
-from .bundle_generate import ComponentsBundle
