+++ csm/core/controllers/__init__.py
@@ -15,10 +15,8 @@
 
 from .routes import CsmRoutes
 from .users import CsmUsersListView, CsmUsersView
-from .alerts.alerts import AlertsView, AlertsListView
-from .alerts.alerts_history import AlertsHistoryListView, AlertsHistoryView
 from .health import ResourcesHealthView
 from .cluster_management import ClusterOperationsView
 from .audit_log import AuditLogShowView, AuditLogDownloadView
 from .maintenance import MaintenanceView
-# from .file_transfer import CsmFileView
+
